Refactor: amd64_x570 interface
