+++ amd_desktop/amd64_x570.py
@@ -6,7 +6,7 @@ import logging
 import math
 import re
 from unit.log_handler import get_logger
-from unit.application_interface import BaseInterface
+from unit.amd64_interface import BaseInterface
 # import os
 
 logger = get_logger(__name__, logging.INFO)
